requirements.txt
asgiref==3.8.1
attrs==23.2.0
cloudinary==1.40.0
dj-database-url==2.1.0
Django==4.0.10
django-cors-headers==4.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.22.1
gunicorn==22.0.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
packaging==24.0
Pillow==9.1.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.0
rpds-py==0.18.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
uritemplate==4.1.1
whitenoise==6.5.0
django-jazzmin
docutils